requirements.txt
kafka-python==1.4.6
pkg-resources==0.0.0
python-dotenv==0.10.3
python-snappy==0.5.4
